@@ -7,7 +7,7 @@ import be.ugent.idlab.knows.functions.agent.functionModelProvider.FunctionModelP
 import be.ugent.idlab.knows.functions.agent.functionModelProvider.fno.FnOFunctionModelProvider;
 import be.ugent.idlab.knows.functions.agent.functionModelProvider.fno.exception.FnOException;
 import be.ugent.idlab.knows.functions.agent.model.Function;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
@@ -15,7 +15,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
-import static org.junit.Assert.*;
+import static org.junit.jupiter.api.Assertions.*;
 
 /**
  * <p>Copyright 2022 IDLab (Ghent University - imec)</p>
@@ -37,7 +37,7 @@ public class InstantiatorTest {
         Object result = sum.invoke(null, 5, 8);
         assertTrue(result instanceof Long);
         Long resultValue = (Long) result;
-        assertEquals("5 + 8 should be 13", 13, resultValue.intValue());
+        assertEquals(13, resultValue.intValue(), "5 + 8 should be 13");
     }
 
     @Test
@@ -45,7 +45,7 @@ public class InstantiatorTest {
         Class<?> clazz = Class.forName("be.ugent.idlab.knows.functions.internalfunctions.InternalTestFunctions");
         Method method = clazz.getMethod("sum", long.class, long.class);
         Object result = method.invoke(null, Long.valueOf(1), Long.valueOf(3));
-        assertEquals("1 + 3 should be 4", 4L, result);
+        assertEquals(4L, result, "1 + 3 should be 4");
         Class<?>[] genericParameterTypes = method.getParameterTypes();
         assertArrayEquals(new Class<?>[]{long.class, long.class}, genericParameterTypes);
     }
@@ -62,6 +62,6 @@ public class InstantiatorTest {
         List anEmptyRawList = new ArrayList();
         Method rawListLen = instantiator.getMethod("http://example.org/rawListLen");
         Object result = rawListLen.invoke(null, anEmptyRawList);
-        assertEquals("An empty list should have size 0", 0L, result);
+        assertEquals(0L, result, "An empty list should have size 0");
     }
 }
