@@ -58,14 +58,12 @@ public class Arguments {
     @Override
     public String toString() {
         final StringBuilder str = new StringBuilder();
-        nameToValueMap.entries().forEach(entry -> {
-            str
-                    .append("('")
-                    .append(entry.getKey())
-                    .append("' -> '")
-                    .append(entry.getValue().toString())
-                    .append("')");
-        });
+        nameToValueMap.entries().forEach(entry -> str
+                .append("('")
+                .append(entry.getKey())
+                .append("' -> '")
+                .append(entry.getValue().toString())
+                .append("')"));
         return str.toString();
     }
 }
